fix: make associated table reusable
